requirements.txt
colorlog
requests
pillow
openai